TINY-4781: rtc integration
