+++ modules/oxide/src/less/theme/content-ui.less
@@ -28,3 +28,4 @@
 @import 'content/table/table';
 @import 'content/visualblocks/visualblocks';
 @import 'content/visualchars/visualchars';
+@import 'content/rtc/rtc';
