+++ modules/tinymce/src/core/main/ts/api/EditorUpload.ts
@@ -12,7 +12,7 @@ import { BlobInfoImagePair, ImageScanner } from '../file/ImageScanner';
 import { Uploader } from '../file/Uploader';
 import UploadStatus from '../file/UploadStatus';
 import Editor from './Editor';
-import { BlobCache } from './file/BlobCache';
+import { BlobCache, BlobInfo } from './file/BlobCache';
 import * as Settings from './Settings';
 
 /**
@@ -25,6 +25,8 @@ import * as Settings from './Settings';
 export interface UploadResult {
   element: HTMLImageElement;
   status: boolean;
+  blobInfo: BlobInfo;
+  uploadUri: string;
 }
 
 export type UploadCallback = (results: UploadResult[]) => void;
@@ -102,17 +104,18 @@ const EditorUpload = function (editor: Editor): EditorUpload {
     });
   };
 
-  const replaceImageUri = function (image: HTMLImageElement, resultUri: string) {
-    blobCache.removeByUri(image.src);
+  const replaceImageUriInView = (image: HTMLImageElement, resultUri: string) => {
+    const src = editor.convertURL(resultUri, 'src');
+
     replaceUrlInUndoStack(image.src, resultUri);
 
     editor.$(image).attr({
       'src': Settings.shouldReuseFileName(editor) ? cacheInvalidator(resultUri) : resultUri,
-      'data-mce-src': editor.convertURL(resultUri, 'src')
+      'data-mce-src': src
     });
   };
 
-  const uploadImages = (callback?: UploadCallback) => {
+  const uploadImages = (callback?: UploadCallback): Promise<UploadResult[]> => {
     if (!uploader) {
       uploader = Uploader(uploadStatus, {
         url: Settings.getImageUploadUrl(editor),
@@ -128,18 +131,22 @@ const EditorUpload = function (editor: Editor): EditorUpload {
       });
 
       return uploader.upload(blobInfos, openNotification).then(aliveGuard((result) => {
-        const filteredResult = Arr.map(result, (uploadInfo, index) => {
+        const filteredResult: UploadResult[] = Arr.map(result, (uploadInfo, index) => {
+          const blobInfo = imageInfos[index].blobInfo;
           const image = imageInfos[index].image;
 
           if (uploadInfo.status && Settings.shouldReplaceBlobUris(editor)) {
-            replaceImageUri(image, uploadInfo.url);
+            blobCache.removeByUri(image.src);
+            replaceImageUriInView(image, uploadInfo.url);
           } else if (uploadInfo.error) {
             ErrorReporter.uploadError(editor, uploadInfo.error);
           }
 
           return {
             element: image,
-            status: uploadInfo.status
+            status: uploadInfo.status,
+            uploadUri: uploadInfo.url,
+            blobInfo,
           };
         });
