+++ modules/tinymce/src/core/main/ts/api/UndoManager.ts
@@ -8,10 +8,10 @@
 import { Event } from '@ephox/dom-globals';
 import { Cell, Option } from '@ephox/katamari';
 import { Bookmark } from '../bookmark/BookmarkTypes';
-import { addUndoLevel, beforeChange, clear, extra, hasRedo, hasUndo, ignore, redo, reset, transact, undo } from '../undo/Operations';
 import { addKeyboardShortcuts, registerEvents } from '../undo/Setup';
 import { UndoManager, Locks, Index, UndoLevel } from '../undo/UndoManagerTypes';
 import Editor from './Editor';
+import * as Rtc from '../Rtc';
 
 /**
  * This class handles the undo/redo history levels for the editor. Since the built-in undo/redo has major drawbacks a custom one was needed.
@@ -42,7 +42,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @method beforeChange
      */
     beforeChange() {
-      beforeChange(editor, locks, beforeBookmark);
+      Rtc.beforeChange(editor, locks, beforeBookmark);
     },
 
     /**
@@ -54,7 +54,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @return {Object} Undo level that got added or null it a level wasn't needed.
      */
     add(level?: UndoLevel, event?: Event): UndoLevel {
-      return addUndoLevel(editor, undoManager, index, locks, beforeBookmark, level, event);
+      return Rtc.addUndoLevel(editor, undoManager, index, locks, beforeBookmark, level, event);
     },
 
     /**
@@ -64,7 +64,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @return {Object} Undo level or null if no undo was performed.
      */
     undo(): UndoLevel {
-      return undo(editor, undoManager, locks, index);
+      return Rtc.undo(editor, undoManager, locks, index);
     },
 
     /**
@@ -74,7 +74,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @return {Object} Redo level or null if no redo was performed.
      */
     redo(): UndoLevel {
-      return redo(editor, index, undoManager.data);
+      return Rtc.redo(editor, index, undoManager.data);
     },
 
     /**
@@ -83,7 +83,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @method clear
      */
     clear() {
-      clear(editor, undoManager, index);
+      Rtc.clear(editor, undoManager, index);
     },
 
     /**
@@ -92,7 +92,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @method reset
      */
     reset() {
-      reset(undoManager);
+      Rtc.reset(editor, undoManager);
     },
 
     /**
@@ -102,7 +102,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @return {Boolean} true/false if the undo manager has any undo levels.
      */
     hasUndo() {
-      return hasUndo(editor, undoManager, index);
+      return Rtc.hasUndo(editor, undoManager, index);
     },
 
     /**
@@ -112,7 +112,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @return {Boolean} true/false if the undo manager has any redo levels.
      */
     hasRedo() {
-      return hasRedo(undoManager, index);
+      return Rtc.hasRedo(editor, undoManager, index);
     },
 
     /**
@@ -126,7 +126,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @return {Object} Undo level that got added or null it a level wasn't needed.
      */
     transact(callback: () => void): UndoLevel {
-      return transact(undoManager, locks, callback);
+      return Rtc.transact(editor, undoManager, locks, callback);
     },
 
     /**
@@ -138,7 +138,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @param {function} callback Function that gets executed and has dom manipulation logic in it.
      */
     ignore(callback: () => void) {
-      ignore(locks, callback);
+      Rtc.ignore(editor, locks, callback);
     },
 
     /**
@@ -151,11 +151,14 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @param {function} callback2 Function that does mutation but gets displayed to the user.
      */
     extra(callback1: () => void, callback2: () => void) {
-      extra(editor, undoManager, index, callback1, callback2);
+      Rtc.extra(editor, undoManager, index, callback1, callback2);
     }
   };
 
-  registerEvents(editor, undoManager, locks);
+  if (!Rtc.isRtc(editor)) {
+    registerEvents(editor, undoManager, locks);
+  }
+
   addKeyboardShortcuts(editor);
 
   return undoManager;
